requirements.txt
absl-py==0.1.13
astor==0.6.2
bleach==1.5.0
bottle==0.12.9
click==6.7
certifi==2017.4.17
chardet==3.0.3
Flask==0.12.2
future==0.16.0
gast==0.2.0
grpcio==1.10.0
gunicorn==19.7.1
h5py==2.7.1
html5lib==0.9999999
itsdangerous==0.24
idna==2.5
Jinja2==2.9.6
Keras==2.1.5
line-bot-sdk==1.5.0
Markdown==2.6.11
MarkupSafe==0.23
numpy==1.14.2
opencv-python==3.4.0.12
oauthlib==2.0.2
pillow==5.0.0
protobuf==3.5.2.post1
PyYAML==3.12
python-dotenv==0.7.1
requests==2.14.2
scipy==1.0.1
six==1.11.0
tensorboard==1.6.0
tensorflow==1.6.0
termcolor==1.1.0
urllib3==1.21.1
Werkzeug==0.14.1

